+++ JDI.Light/JDI.Light.Tests/Tests/Complex/CheckListTests.cs
@@ -1,6 +1,9 @@
 ﻿using System.Collections.Generic;
+using JDI.Light.Elements.Common;
 using JDI.Light.Exceptions;
+using JDI.Light.Interfaces.Common;
 using JDI.Light.Interfaces.Complex;
+using JDI.Light.Interfaces.Complex.Generic;
 using NUnit.Framework;
 
 namespace JDI.Light.Tests.Tests.Complex
@@ -10,7 +13,11 @@ namespace JDI.Light.Tests.Tests.Complex
     {
         private readonly string text = "Hot option";
 
-        private ICheckList _weather;
+        //private ICheckList _weather;
+
+        //private ICheckList<ICheckBox> _weather;
+
+        private ICheckList<MyCheckBox> _weather;
 
         [SetUp]
         public void SetUp()
@@ -18,7 +25,8 @@ namespace JDI.Light.Tests.Tests.Complex
             TestSite.Html5Page.Open();
             TestSite.Html5Page.CheckOpened();
             TestSite.Html5Page.WeatherCheckList.Check(text);
-            _weather = TestSite.Html5Page.WeatherCheckList;
+            //_weather = TestSite.Html5Page.WeatherCheckList;
+            _weather = TestSite.Html5Page.WeatherMyCheckbox;
         }
         
         [Test]
