+++ JDI.Light/JDI.Light/Interfaces/Complex/Generic/ICheckListcs.cs
@@ -1,0 +1,26 @@
+﻿using System.Collections.Generic;
+using JDI.Light.Interfaces.Base;
+using JDI.Light.Interfaces.Common;
+using OpenQA.Selenium;
+
+namespace JDI.Light.Interfaces.Complex.Generic
+{
+    public interface ICheckList<TCHeckBox> : IBaseUIElement, IGetValue<List<string>> where TCHeckBox : ICheckBox
+    {
+        By CheckListLocator { get; set; }
+        By LabelLocator { get; set; }
+        void Check(params string[] values);
+        void Check(params int[] indexes);
+        void Uncheck(params string[] values);
+        void Uncheck(params int[] indexes);
+        void Select(params string[] values);
+        void Select(params int[] indexes);
+        void UncheckAll();
+        void CheckAll();
+        string[] Checked();
+        bool IsChecked(string value);
+        bool IsChecked(int index);
+        bool IsDisabled(string value);
+        bool IsDisabled(int index);
+    }
+}
