+++ JDI.Light/JDI.Light/JDI.Light.csproj
@@ -74,6 +74,7 @@
     <Compile Include="Elements\Common\DateTimeSelector.cs" />
     <Compile Include="Elements\Common\NumberSelector.cs" />
     <Compile Include="Elements\Common\ProgressBar.cs" />
+    <Compile Include="Elements\Complex\Generic\CheckList.cs" />
     <Compile Include="Elements\Complex\RadioButtons.cs" />
     <Compile Include="Elements\Common\Range.cs" />
     <Compile Include="Elements\Complex\Table\Column.cs" />
@@ -115,6 +116,7 @@
     <Compile Include="Interfaces\Common\IButton.cs" />
     <Compile Include="Interfaces\Common\ICheckBox.cs" />
     <Compile Include="Interfaces\Common\IColorPicker.cs" />
+    <Compile Include="Interfaces\Complex\Generic\ICheckListcs.cs" />
     <Compile Include="Interfaces\Complex\ICheckList.cs" />
     <Compile Include="Interfaces\Common\IDataList.cs" />
     <Compile Include="Interfaces\Common\IDatePicker.cs" />
